requirements.txt
Flask==0.10.1
webargs==1.0.0
SQLAlchemy==1.0.9
python-instagram==1.3.2